+++ src/CodeGeneration.Roslyn.Tests.Generators/AddExampleBuildPropertyAttribute.cs
@@ -1,0 +1,16 @@
+// Copyright (c) Andrew Arnott. All rights reserved.
+// Licensed under the MS-PL license. See LICENSE.txt file in the project root for full license information.
+
+namespace CodeGeneration.Roslyn.Tests.Generators
+{
+    using System;
+    using System.Diagnostics;
+    using Validation;
+
+    [AttributeUsage(AttributeTargets.Class, Inherited = false, AllowMultiple = true)]
+    [CodeGenerationAttribute(typeof(AddExampleBuildPropertyGenerator))]
+    [Conditional("CodeGeneration")]
+    public class AddExampleBuildPropertyAttribute : Attribute
+    {
+    }
+}